requirements.txt
absl-py
attrs
click
gymnasium
idna
imageio
matplotlib
mpyq
numpy
pbr
Pillow
pluggy
portpicker
probscale
protobuf
py
pygame
pyparsing
pytest
python-dateutil
PyYAML==5.3.1
requests
sacred
seaborn
scipy
six
sk-video
snakeviz
tensorboard-logger
torch
torchvision
tornado
urllib3
websocket-client
whichcraft
wrapt
ml-collections
